Update variable type in 01_variables.c and add arithmetic operations in 02_arithmetic_operators.c and type conversion examples in 03_type_conversion.c

+++ Chapter_02/learning_02/01_variables.c
@@ -2,10 +2,10 @@
 
 int main() {
     int i = 10, p = 2;
-    int j = i;
+    float j = i;
     int a, b, c, d;
     a = b = c = d = 20;
-    printf("The value of i is %d and value of j is %d\n", i, j);
+    printf("The value of i is %d and value of j is %.2f\n", i, j);
     int j1 = p + j - i;
     printf("The value of j1 is %d\n", j1);
     printf("The value of a is %d, b is %d, c is %d, and d is %d\n", a, b, c, d);
